backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.10
pydantic-settings==2.5.2
email-validator==2.2.0
boto3==1.35.27
openai==1.47.0
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
python-dotenv==1.0.1
httpx==0.27.2
reportlab==4.2.2
